+++ packages/web-components/keen-slider-autoplay/src/components/index.ts
@@ -1,0 +1,1 @@
+export * as KeenSliderAutoplay from "./keen-slider-autoplay.js";