+++ packages/web-components/keen-slider-navigation-dots/src/components/keen-slider-navigation-dots.ts
@@ -94,7 +94,7 @@ export const Component: CorrectComponentType<Props> = (props, { element }) => {
     },
     () => null,
     ([_, { addPlugin }]) => {
-      addPlugin((slider) => {
+      const removePlugin = addPlugin((slider) => {
         setSlider(slider);
 
         slider.on("created", function (slider) {
@@ -108,6 +108,8 @@ export const Component: CorrectComponentType<Props> = (props, { element }) => {
           }),
         );
       });
+
+      onCleanup(removePlugin);
     },
   );